feat(benchmarks): update mega benchmark to measure execution time and add C# LINQ benchmark feat(ast): change int_value type to long long for better integer handling feat(te_bytecode): enhance JIT support for Windows and improve type detection logic

+++ src/ast.h
@@ -137,7 +137,7 @@ typedef struct Variable {
     ValueType vtype;
     char *type;
     union {
-        int int_value;
+        long long int_value;
         char *string_value;
         double float_value;
         struct ObjectNode *object_value;
